+++ app/scripts/views/page2/Page2Ctrl.ts
@@ -1,18 +1,13 @@
 // Created by baihuibo on 16/3/24.
 
-import {Controller, Menu, Route} from "annotation";
+import {Controller, Route} from "annotation";
 import {IScope} from "../../typings/common/IScope";
 
-let name = "页面2";
-let icon = "icon";
-let route = "/manage2";
-
 @Route({
-    route: route,
+    route: "/page2",
     controllerAs: "T",
     templateUrl: "scripts/views/page2/page.html"
 })
-@Menu({icon, name, route})
 @Controller
 export class Page2Ctrl {
     public name:string;