+++ changelog.md
@@ -1,3 +1,29 @@
+# v0.0.45
+- 新增
+    1. `js-yaml` 解析模块
+    2. 新增菜单配置文件 `app/data/menu.yaml`
+    
+- 变更
+    1. 取消注解 Menu 已不需要
+    2. ctrl 中只需参数 `route` 即可
+       ```ts
+       import {Controller, Route} from "annotation";
+       import {IScope} from "../../typings/common/IScope";
+       @Route({
+           route: "/page2", //只需要此route参数,之前的参数 name和icon之类的需写在配置文件[menu.yaml]中
+           controllerAs: "T",
+           templateUrl: "scripts/views/page2/page.html"
+       })
+       @Controller
+       export class Page2Ctrl {
+           public name:string;
+       
+           constructor(public $scope:IScope) {
+               this.name = 'page2';
+           }
+       }
+       ```
+
 # v0.0.44
 - 新增
     1. 分页指令新增数据格式化参数 `format` 用来做数据格式化处理操作